Created separate functions to add the quadratic costs.
* Each solver class is provided with a function to add the quadratic
cost. The function takes in the decision variable vectors and quadratic
cost matrices. The matrices are optional and default to the identity
matrix.

* Each of the previous constructors now has a counterpart that takes in
the vectors and matrices. Vectors are not passed to the original
constructor. These constructors don't setup the cost, requiring the user
to do so.

* The test file has been updated to reflect the changes. More
constructor tests have been added.